Add logs, refactor session/shell
