+++ src/windows/wslc/services/SessionService.h
@@ -17,8 +17,17 @@ Abstract:
 #include <wslaservice.h>
 
 namespace wsl::windows::wslc::services {
+struct SessionInformation
+{
+    ULONG SessionId;
+    DWORD CreatorPid;
+    std::wstring DisplayName;
+};
+
 struct SessionService
 {
+    static int Attach(const std::wstring& name);
     static wsl::windows::wslc::models::Session CreateSession(const std::optional<wsl::windows::wslc::models::SessionOptions>& options = std::nullopt);
+    static std::vector<SessionInformation> List();
 };
 } // namespace wsl::windows::wslc::services
