--- src/windows/wslc/services/ShellService.h
@@ -1,32 +1,0 @@
-/*++
-
-Copyright (c) Microsoft. All rights reserved.
-
-Module Name:
-
-    ShellService.h
-
-Abstract:
-
-    This file contains the ShellService definition
-
---*/
-#pragma once
-
-#include <string>
-#include <vector>
-
-namespace wsl::windows::wslc::services {
-struct SessionInformation
-{
-    ULONG SessionId;
-    DWORD CreatorPid;
-    std::wstring DisplayName;
-};
-
-struct ShellService
-{
-    static int Attach(const std::wstring& name);
-    static std::vector<SessionInformation> List();
-};
-} // namespace wsl::windows::wslc::services
