+++ src/windows/wslc/tasks/ContainerTasks.h
@@ -18,7 +18,6 @@ using wsl::windows::wslc::execution::CLIExecutionContext;
 
 namespace wsl::windows::wslc::task {
 void CreateContainer(CLIExecutionContext& context);
-void CreateSession(CLIExecutionContext& context);
 void DeleteContainers(CLIExecutionContext& context);
 void ExecContainer(CLIExecutionContext& context);
 void GetContainers(CLIExecutionContext& context);
@@ -29,4 +28,5 @@ void RunContainer(CLIExecutionContext& context);
 void SetContainerOptionsFromArgs(CLIExecutionContext& context);
 void StartContainer(CLIExecutionContext& context);
 void StopContainers(CLIExecutionContext& context);
+void ViewContainerLogs(CLIExecutionContext& context);
 } // namespace wsl::windows::wslc::task
