+++ src/windows/wslc/tasks/SessionTasks.h
@@ -17,6 +17,7 @@ Abstract:
 using wsl::windows::wslc::execution::CLIExecutionContext;
 
 namespace wsl::windows::wslc::task {
-void ListSessions(CLIExecutionContext& context);
 void AttachToSession(CLIExecutionContext& context);
+void CreateSession(CLIExecutionContext& context);
+void ListSessions(CLIExecutionContext& context);
 } // namespace wsl::windows::wslc::task
